Don't fcd_conf_disk_count in fcd_conf_disk_idx()
We need to ensure that the [freecusd] config file section has been
processed before any options in [raid_disk:*] sections.  We
currently check the value of fcd_conf_disk_count, which is
initialized to zero and then set when [freecusd]:raid_disks is
processed (possibly by setting it to its default value).  Once the
RAID disks are auto-detected, fcd_conf_disk_count will be set by
the auto-detection code, so it won't be usable for this purpose.

Instead, initialize fcd_conf_disks[0].temp_warn to INT_MIN and
check for that value.  (It will be overwritten when
[freecusd]:hdd_temp_warn is processed.)

Also unify checking of HDD temperature values and prohibit
temperatures less than absolute zero.